add tests for moveBootstrapFunction
