+++ lib/v4/migration-helpers/update-plugin-folder-structure/utils/create-content-type-index.js
@@ -1,0 +1,26 @@
+const { join } = require('path');
+const fs = require('fs-extra');
+
+const { importFilesToIndex, addModulesToExport } = require('../transforms');
+
+const createDirectoryIndex = require('./create-directory-index');
+
+async function createContentTypeIndex(v4PluginPath, dir) {
+  const hasDir = await fs.pathExists(dir);
+  if (!hasDir) return;
+
+  const indexPath = join(dir, 'index.js');
+
+  await fs.copy(join(__dirname, '..', 'utils', 'module-exports.js'), indexPath);
+  const dirContent = await fs.readdir(dir, { withFileTypes: true });
+  const directoriesToImport = dirContent.filter((fd) => fd.isDirectory()).map((fd) => fd.name);
+
+  for (dir of directoriesToImport) {
+    createDirectoryIndex(join(v4PluginPath, 'server', 'content-types', dir));
+  }
+
+  await importFilesToIndex(indexPath, directoriesToImport);
+  await addModulesToExport(indexPath, directoriesToImport);
+}
+
+module.exports = createContentTypeIndex;
